C2/c2_globo/c2_globo/settings/heroku.py
import environ

from c2_globo.settings.base import *

env = environ.Env()

Debug = env.bool("DEBUG", False)

SECRET_KEY = env("SECRET_KEY")

ALLOWED_HOSTS = env.list("ALLOWED_HOSTS")

DATABASES = {
    "default": env.db(),
}